Add initial realm commands

+++ src/realm/ConnectionStats.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 Ember
+ * Copyright (c) 2016 - 2026 Ember
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -9,6 +9,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 
 namespace ember::realm {
 
@@ -19,7 +20,7 @@ struct ConnectionStats {
 	std::size_t messages_out;
 	std::size_t async_receives;
 	std::size_t async_sends;
-	std::size_t latency;
+	std::uint32_t latency;
 };
 
 } // realm, ember
